WrapsResult: Move off duplicate code into a trait
